Correct absent default state value in ui/dark-mode
The default value for the `theme` state variable, assuming that a value hasn't
been persisted to local storage, is now `auto`. This prevents an error throw in
nav/menu-color, which is the only component to rely on `theme` and not
`effectiveTheme` and therefore the first component to run afoul when `theme` is
null.

Again, the issue corrected here can only be detected when local storage doesn't
contain theme info. This issue was created in 485aba5 when the nav/menu-color
was added. Before finalizing that commit, the issue was noticed, but
misunderstood and incorrectly fixed. After the issue appeared to vanish, the fix
was removed and never committed.