Add Netlify config to fix 404 error
Previously, when the user tried to navigate to other pages in the app by
changing the URL, a 404 error would be displayed. This would happened on the
prod environment, but not locally.

In this fix, I applied a Netlify redirects file to makes sure all index page
will be loading if the user directly changes the URL.

See Netlify docs link below for info relavent info for SPA:
https://docs.netlify.com/routing/redirects/rewrites-proxies/#history-pushstate-and-single-page-apps