Restore previous seed data, fix typos in init_db
For unknown reasons, the seed data for the auth.users table was altered while
finishing commit f3b8c08. The alteration appears to be very slight whitespace
differences, but it caused `seed_db.sql` to be rejected because of a syntax
error in the uuid type column. By revisiting PR #12 (which was shelved because
it was too large), I found a working seed statement and included it in this
commit.

In that same problematic commit, there were typos in `init_db.sh` with the env
variables. Those typos broke the script and illustrated that I failed to test my
work before committing it.